Fix basic message generation issues
There were a few simple issues with generating messages from a FIXMessage
object. An existing field 10 would be written, rather than letting the new
message's checksum be written at the end. A separator was written twice in
one place. And finally, the checksum itself was written with space padding.
Now all of these issues are fixed.